feat: confirm users want to leave during restart
When setting up the pi-top or waiting for the reboot to complete it is
potentially destructive or confusing to reload or close the browser tab.

Intercept the 'beforeunload' event and prevent the unload while either
of the above are happening.

Defining a custom message for the confirmation dialog is no longer
possible with modern browsers: https://stackoverflow.com/a/50132210/5533971