chore: add msw
Add msw to support user-centric testing and better development
experience when not running on the pi.

Add msw folder to src. CRA does not support importing from files outside
of src.

Add worker and rest helpers to the browser window in development mode to
allow modifying the default handlers at runtime.

Ensure the frontend uses relative endpoints in development by default so
that the msw worker can intercept those requests. To bypass the worker
and use a development version of the backend set REACT_APP_API_PORT to
match the port it is running on.