feat: add rover controller to landing page
Create a systemd service for the rover controller. Using systemd gives
us an API for starting and stopping the server while keeping the
controller decoupled from the web portal.

Provide routes on the backend for enabling, disabling and getting the
status of the systemd service. The systemd service status can be
'active' while the controller server is still starting up so verify the
server is ready before responding with an active status.

Add RoverControllerLanding component which can launch and stop the
controller. Show loading states when launching and stopping the
controller and provide troublshooting when the controller crashes.

Modify the tsconfig to allow implicit returns since they are common in
effects that conditionally need to cleanup as can be seen in the
RoverControllerLanding component.

When running on the web renderer we want to give the user a warning that
they should use the rover controller on a separate device to the pi-top.
This is because the landing page is only shown after onboarding at which
point they are either using a screen and keyboard, which can't be used
to remotely control the pi-top, or VNC which would degrade the
experience significantly.