feat: show error dialog when disconnected from AP
The existing error is quite subtle and looks like it is an issue with
the password since it is just below the password input. It is also
possible to escape the dialog and continue without a connection to the
pi-top, this will cause unexpected issues later.

Replace the error message with a dialog that explains what to do in more
detail and prevents the user from continuing with onboarding until they
are reconnected to the pi-top.

Preload the connect-to-wifi image since it is not possible to load the
image after the connection to AP is lost.