chore: remove duplicated server and client routing
Right now we are duplicating route matching on server and the client,
this means that any changes to the routes need to be matched on both
sides which is easy to miss or forget.

Remove the FrontendRoutes enum and corresponding is_valid class method
since that validation can be deferred to the client-side. The only time
we need server-side routing is when we need to redirect users opening
pi-top.local to /onboarding if it has not yet been completed.

We don't need to send users to /onboarding for every route if it has not
been completed.

Update client-side routing to fallback to Landing page to match the
existing routing behaviour.

A nice side-effect of this is that running the frontend without the
server doesn't show a blank screen when opened anymore.