fix: remove wifi and updater page close buttons
When showing the standalone wifi and updater pages using the
web-renderer we show custom close buttons that hit an endpoint to close
the window. This was done in the past since the window frame was not
being shown and so there was no close button rendered. The wifi and
updater pages are being shown with the window frame visible however, so
we don't need the custom close buttons.