fix: hide finish button in standalone updater
The Finish button in the standalone updater does not work when using the
web-renderer, this might be a backend problem. We don't actually want
users to click the Finish button anyways though as the window has it's
own close button.

Hide the finish button when standlone and update the messages to
indicate the window can be closed by the user.