fix: onboarding on mobile devices
- fix onboarding buttons hidden on Android

Due to the url bar on Android taking up some of the screen but vh being
1% of the window height including the url bar the bottom buttons on the
onboarding app are obscured.

Add -webkit-fill-available as a max-height to fix the height on Android
while using 100vh still on unsupported devices.

- fix dialog width being too wide for mobiles

The dialog default width is wider than some mobile devices, this causes
the action buttons to be off the screen.

Ensure the width of the Dialog and the actions passed in responsively
handle small width screens.

- fix dialog too tall on mobile landscape

The Dialog height is greater than the height of mobiles on landscape,
use a min modifier to ensure the height of the dialog is always less
than 90% of the screen height.

- fix build info intercepting skip button clicks

On mobile devices in portrait mode the build info overlaps with the skip
button, this prevents users from clicking skip.

Hide the build info on portrait mobile devices. Show the build info if
the user is in landscape mode to allow users to easily access it for
support purposes.

Move the build info below the app contents so that if it accidentally
overlaps with another element it won't intercept clicks.

- fix buttons offscreen on landscape mobile

When on landscape mobile the next and skip buttons are being pushed
offscreen. Hide the banner image to allow the content enough space when
on landscape mobile screens.